[package] Pull out _UnpicklerWrapper into PackageUnpickler
This makes our API symmetric--now we have an `Importer` aware Pickler
and Unpickler implementation that have similar interfaces.

[ghstack-poisoned]
