+++ torch/package/package_importer.py
@@ -5,10 +5,8 @@ import inspect
 import io
 import linecache
 from weakref import WeakValueDictionary
-import pickle
 import torch
 from torch.serialization import _get_restore_location, _maybe_decode_ascii
-import _compat_pickle  # type: ignore
 import types
 import os.path
 from pathlib import Path
@@ -18,6 +16,7 @@ from ._importlib import _normalize_line_endings, _resolve_name, _sanity_check, _
 from ._file_structure_representation import _create_folder_from_file_list, Folder
 from ._glob_group import GlobPattern
 from ._mock_zipreader import MockZipReader
+from ._package_unpickler import PackageUnpickler
 from ._mangling import PackageMangler, demangle
 from .importer import Importer
 
@@ -91,7 +90,7 @@ class PackageImporter(Importer):
         self._mangler = PackageMangler()
 
         # used for torch.serialization._load
-        self.Unpickler = lambda *args, **kwargs: _UnpicklerWrapper(self.import_module, *args, **kwargs)
+        self.Unpickler = lambda *args, **kwargs: PackageUnpickler(self, *args, **kwargs)
 
     def import_module(self, name: str, package=None):
         """Load a module from the package if it hasn't already been loaded, and then return
@@ -199,11 +198,11 @@ class PackageImporter(Importer):
         return self._mangler.parent_name()
 
     def file_structure(self, *, include: 'GlobPattern' = "**", exclude: 'GlobPattern' = ()) -> Folder:
-        """Returns a file structure representation of package's zipfile. 
+        """Returns a file structure representation of package's zipfile.
 
         Args:
             include (Union[List[str], str]): An optional string e.g. "my_package.my_subpackage", or optional list of strings
-                for the names of the files to be inluded in the zipfile representation. This can also be 
+                for the names of the files to be inluded in the zipfile representation. This can also be
                 a glob-style pattern, as described in exporter's :meth:`mock`
 
             exclude (Union[List[str], str]): An optional pattern that excludes files whose name match the pattern.
@@ -459,21 +458,6 @@ class PackageImporter(Importer):
 _NEEDS_LOADING = object()
 _ERR_MSG_PREFIX = 'No module named '
 _ERR_MSG = _ERR_MSG_PREFIX + '{!r}'
-
-class _UnpicklerWrapper(pickle._Unpickler):  # type: ignore
-    def __init__(self, importer, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._importer = importer
-
-    def find_class(self, module, name):
-        # Subclasses may override this.
-        if self.proto < 3 and self.fix_imports:
-            if (module, name) in _compat_pickle.NAME_MAPPING:
-                module, name = _compat_pickle.NAME_MAPPING[(module, name)]
-            elif module in _compat_pickle.IMPORT_MAPPING:
-                module = _compat_pickle.IMPORT_MAPPING[module]
-        mod = self._importer(module)
-        return getattr(mod, name)
 
 class _PathNode:
     pass
